// teste-courart-frontend/src/app/model/veiculo.ts
export class Veiculo {
    codigo: number;
    placa: string;
    ativo: boolean;
    anoFabricacao: string;
    anoModelo: string;
    chassi: string;
    dataCadastro: string;
    dataDesativacao: string;
    modelo: string;
    cor: string;
    qtdPassageiros: number;
    consumoMedio: number;
}